修正BUG

+++ foreign_redis_procedure/pgsql.sql
@@ -28,7 +28,7 @@ CREATE TABLE base_redis_foreigns (
 	dbindex SMALLINT DEFAULT 0,
 	"method" methods DEFAULT 'get',
 	"using" VARCHAR(100) DEFAULT NULL,
-	restful VARCHAR(200) DEFAULT NULL,
+	"restful" VARCHAR(200) DEFAULT NULL,
 	route JSON DEFAULT NULL,
 	datetime TIMESTAMP(0) WITHOUT TIME ZONE NOT NULL DEFAULT now(),
 	PRIMARY KEY (key)
@@ -53,125 +53,6 @@ COMMENT ON COLUMN base_redis_foreigns.datetime IS '最后更新时间';
 
 /**
  # PostgreSQL redis操作存储过程
-
-## 1. 依耐关系(首先需要安装`Redis_fdw`)
-> 关于Redis_fdw 详见：https://pgxn.org/dist/redis_fdw/
-```sql
-CREATE EXTENSION IF NOT EXISTS redis_fdw;
-CREATE SERVER redis_server FOREIGN DATA WRAPPER redis_fdw OPTIONS (address '127.0.0.1', port '6379');
-CREATE USER MAPPING FOR PUBLIC SERVER redis_server OPTIONS (password '');
-```
-
-## 2. 操作方法、使用范列
-> * `structure_redis` 存储过程可以用于某个表的触发器中,或其他数据逻辑操作的存储过程中
-> * 当数据层某个表的数据发生变化时运行该存储过程快速创建、更新或删除Redis数据，应用层不需要去操作`Redis`，只需要读取`Redis`即可
-> * 会在数据库中生成一个`"public"."police_redis_foreign"`的数据表,用于存放Redis服务器数据库MAP(0~15)，请自行规划好Redis数据存放位置
-> * `Redis`服务器数据库15位置处，`KEY` 值为`indexedDBStorage:redisForeign`存放着`"public"."base_redis_foreign"`的数据表的数据，方便应用层查找调用数据
-> * `Redis`服务器数据库15位置处，`KEY` 值为`indexedDBStorage:redisMap`存放着`"public"."base_redis_map_view"`的视图的数据，方便应用层查找调用数据
-> * `Redis`的`KEY`值生成规则：`storage:pk`
-### 2.1 清空`Redis`缓存
-```sql
-SELECT structure_redis(json_build_object(
-	'type', 'empty',
-	'foreign', '外部表名称(不指定则清空所有)',
-	'redis', 'base_redis(默认)'
-);
-```
-
-### 2.2. 创建或刷新一条`JSON`对象形式的`Redis`缓存记录，用于单一数据的缓存
-*  #### 2.2.1 将指定表`table`的`where`查询结果的第一行记录以JSON对象形式进行Redis数据缓存
-```sql
-SELECT structure_redis(
-	json_build_object(
-		'pk', 4,                                    -- 数据主键值
-		'type', 'info',															-- 数据类型, 必须指定为`info`
-		'databases', 1,															-- 指定Redis的databases，范围：0~15， 默认：0
-		'storage', 'sessionStorage',                -- 前端存储位置, 用于构造Redis的Key，默认indexedDBStorage
-		'store', 'userInfo',                        -- 前端存储store, 用于构造Redis的Key
-		'table', 'police_view_user',                -- 数据查询表名
-		'where', 'uid=4'                            -- 数据查询条件
-	)
-);
-```
-* #### 2.2.2 将指定表`table`的主键`primary`等于`pk`的查询结果的第一行记录以JSON对象形式进行Redis数据缓存
-```sql
-SELECT structure_redis(
-	json_build_object(
-		'pk', 4,                                    -- 数据主键值
-		'type', 'info',															-- 数据类型, 必须指定为`info`
-		'databases', 1,															-- 指定Redis的databases，范围：0~15， 默认：0
-		'storage', 'sessionStorage',                -- 前端存储位置, 用于构造Redis的Key，默认indexedDBStorage
-		'store', 'userInfo',                        -- 前端存储store, 用于构造Redis的Key
-		'table', 'police_view_user',                -- 数据查询表名
-		'primary', 'uid'                            -- 查询主键列名
-	)
-);
-```
-### 2.3 将指定表`table`的`where`查询结果以JSON数组对象形式创建或刷新一条Redis缓存记录，用于数据列表的缓存
-```sql
-SELECT structure_redis(
-	json_build_object(
-		'pk', 'policerList',                            -- 数据PK值(此处并非主键值，用以标注唯一性)
-		'type', 'list',																	-- 数据类型，可以不必指定，如果指定必须指定为`list`
-		'databases', 1,																	-- 指定Redis的databases，范围：0~15， 默认：0
-		'storage', 'indexedDB',                  -- 前端存储位置, 用于构造Redis的Key，默认indexedDBStorage
-		'store', 'userList',                            -- 前端存储store, 用于构造Redis的Key
-		'table', 'police_view_user_list',               -- 数据查询表名
-		'where', 'user_group::JSONB @> json_build_array(''police'')::JSONB'    -- 数据查询条件
-	)
-);
-```
-
-### 2.4 将指定数据`data`进行Redis数据缓存
-```sql
-SELECT base_structure_redis(
-				 json_build_object(
-					 'schemas', 'ionic',
-					 'pk', 'policerList',															-- 数据PK值(此处并非主键值，用以标注唯一性)
-					 'type', 'list',																	-- 数据类型，可以不必指定，如果指定必须指定为`list`
-					 'dbindex', 1,																		-- 指定Redis的databases，范围：0~15， 默认：0
-					 'storage', 'indexedDB',													-- 前端存储位置, 用于构造Redis的Key，默认indexedDB
-					 'store', 'user',																	-- 前端存储store, 用于构造Redis的Key
-					 'table', 'police_view_user_list',								-- 数据查询表名
-					 'where', 'user_group::JSONB @> json_build_array(''police'')::JSONB',		-- 数据查询条件
-					 'restful', 'api/base/getUserList'
-						 )
-					 );
-```
-
-### 2.5 删除一条指定`pk`的Redis缓存记录
-```sql
-SELECT structure_redis(
-	json_build_object(
-		'pk', 4,
-		'schemas', 'ionic',
-		'storage', 'indexedDB',
-		'store', 'user'
-	)
-);
-```
-或者
-```sql
-SELECT base_structure_redis(json_build_object('action', 'remove', 'key', 'ionic.indexedDB_user:19'));
-```
-
-## 3 参数说明
-
-参数|参数说明|可选值|类型|是否必填|默认值
--|-|-|-|-|-
-type|数据或操作类型|list、info、remove、empty|ENUM|false|list
-schemas|缓存分区，可用于按用户组或项目功能进行数据可见性的分区|NULL|string|false|app
-storage|前端本地存储方式|NULL|string|除完全清空Redis外均必填|NULL
-store|前端本地存储store名|NULL|string|除完全清空Redis外均必填|NULL
-table|缓存所需的数据库表名|NULL|string|缓存数据时未指定data情况下必填|NULL
-where|缓存所需的数据库查询方法|NULL|string|false|NULL
-data|所需缓存的数据|NULL|JSON|缓存数据时未指定table情况下必填|NULL
-dbindex|Redis数据库ID，该缓存数据在Redis中的位置|NULL|number|false|0
-pk|当缓存列表数据时为list；当缓存info数据时为数据主键|NULL|string or number|缓存数据类型为info时必填|list
-method|数据请求方法|get、post、put、delete、options|ENUM|false|get
-restful|RestFul请求路径|NULL|string|新增缓存数据时必填|NULL
-route|数据Route请求方式所需的参数|NULL|JSON|false|NULL
-
  */
 CREATE OR REPLACE FUNCTION structure_redis(
 	IN redis JSON
@@ -286,7 +167,6 @@ BEGIN
 	CASE actions
 		WHEN 'rebuild' THEN
 		PERFORM structure_redis(json_build_object('action', 'empty', 'schemas', 'all'));
-		DELETE FROM base_redis_foreigns WHERE "table" IS NULL;
 		FOR redisRecord IN SELECT * FROM base_redis_foreigns LOOP
 			CASE redisRecord.type
 				WHEN 'info' THEN
@@ -299,11 +179,11 @@ BEGIN
 							        'schemas', redisRecord.schemas,
 							        'storage', redisRecord.storage,
 							        'store', redisRecord.store,
-							        'type', redisRecord.type,
+							        'type', redisRecord."type",
 							        'dbindex', redisRecord.dbindex,
-							        'method', redisRecord.method,
+							        'method', redisRecord."method",
 							        'using', redisRecord."using",
-							        'restful', redisRecord.restful,
+							        'restful', redisRecord."restful",
 							        'route', redisRecord.route,
 							        'table', redisRecord."table",
 							        'primary', redisRecord."primary"
@@ -314,25 +194,25 @@ BEGIN
 				PERFORM structure_redis(
 					        json_build_object(
 						        'key', redisRecord.key,
-						        'type', redisRecord.type,
+						        'type', redisRecord."type",
 						        'dbindex', redisRecord.dbindex,
-						        'method', redisRecord.method,
+						        'method', redisRecord."method",
 						        'using', redisRecord."using",
-						        'restful', redisRecord.restful,
+						        'restful', redisRecord."restful",
 						        'route', redisRecord.route,
 						        'table', redisRecord."table",
-						        'where', redisRecord.where
+						        'where', redisRecord."where"
 							        )
 						        );
 			ELSE
 				PERFORM structure_redis(
 					        json_build_object(
 						        'key', redisRecord.key,
-						        'type', redisRecord.type,
+						        'type', redisRecord."type",
 						        'dbindex', redisRecord.dbindex,
-						        'method', redisRecord.method,
+						        'method', redisRecord."method",
 						        'using', redisRecord."using",
-						        'restful', redisRecord.restful,
+						        'restful', redisRecord."restful",
 						        'route', redisRecord.route,
 						        'data', redisRecord."data"
 							        )
@@ -364,13 +244,15 @@ BEGIN
 				SELECT "key", "foreigns", "schemas" INTO item FROM base_redis_foreigns WHERE "foreigns" = foreign_table;
 				executesql := 'DELETE FROM ' || quote_ident(item.schemas || '_redis_indexedDB_foreigns') || ' WHERE key = $1;';
 				EXECUTE executesql USING item.key;
-				executesql := 'DELETE FROM ' || quote_ident(foreign_table) || ';';
-				EXECUTE executesql;
-				executesql := 'DROP FOREIGN TABLE IF EXISTS ' || quote_ident(foreign_table)|| ';';
-				EXECUTE executesql;
+				IF (foreign_table <> item.schemas || '_redis_indexedDB_foreigns') THEN
+					executesql := 'DELETE FROM ' || quote_ident(foreign_table) || ';';
+					EXECUTE executesql;
+					executesql := 'DROP FOREIGN TABLE IF EXISTS ' || quote_ident(foreign_table)|| ';';
+					EXECUTE executesql;
+				END IF;
 			END LOOP;
 			FOR item IN SELECT DISTINCT "schemas" FROM base_redis_foreigns LOOP
-				executesql := 'DROP FOREIGN TABLE IF EXISTS ' || quote_ident(item.schemas || ' _redis_indexedDB_foreigns') || ';';
+				executesql := 'DROP FOREIGN TABLE IF EXISTS ' || quote_ident(item.schemas || '_redis_indexedDB_foreigns') || ';';
 				EXECUTE executesql;
 			END LOOP;
 		END IF;
@@ -432,10 +314,10 @@ BEGIN
 				filed_array := array_append(filed_array, 'dbindex');
 			END IF;
 			IF (json_extract_path_text(redis, 'method') IS NULL) THEN
-				filed_array := array_append(filed_array, 'method');
+				filed_array := array_append(filed_array, '"method"');
 			END IF;
 			IF (json_extract_path_text(redis, 'restful') IS NULL) THEN
-				filed_array := array_append(filed_array, 'restful');
+				filed_array := array_append(filed_array, '"restful"');
 			END IF;
 			IF (json_extract_path_text(redis, 'route') IS NULL) THEN
 				filed_array := array_append(filed_array, 'route');
@@ -448,18 +330,28 @@ BEGIN
 				EXECUTE executesql INTO redisRecord USING key_val;
 				IF (json_extract_path_text(redis, 'dbindex') IS NULL) THEN
 					dbindex := redisRecord.dbindex;
+				ELSE
+					dbindex := json_extract_path_text(redis, 'dbindex');
 				END IF;
 				IF (json_extract_path_text(redis, 'method') IS NULL) THEN
-					method_val := redisRecord.method;
+					method_val := redisRecord."method";
+				ELSE
+					method_val := json_extract_path_text(redis, 'method');
 				END IF;
 				IF (json_extract_path_text(redis, 'restful') IS NULL) THEN
-					restful_val := redisRecord.restful;
+					restful_val := redisRecord."restful";
+				ELSE
+					restful_val := json_extract_path_text(redis, 'restful');
 				END IF;
 				IF (json_extract_path_text(redis, 'route') IS NULL) THEN
 					route_val := redisRecord.route;
+				ELSE
+					route_val := json_extract_path_text(redis, 'route');
 				END IF;
 				IF (json_extract_path_text(redis, 'using') IS NULL) THEN
 					using_val := redisRecord."using";
+				ELSE
+					using_val := json_extract_path_text(redis, 'using');
 				END IF;
 			END IF;
 		END IF;
@@ -489,6 +381,7 @@ BEGIN
 				END IF;
 			END CASE;
 		ELSE
+			type_val = 'custom';
 			custom_val := data_val;
 		END IF;
 		cache_val := json_build_object(
@@ -504,7 +397,7 @@ BEGIN
 		ELSE
 			where_val := NULL;
 		end if;
-		executesql := 'INSERT INTO public.base_redis_foreigns ("key", "foreigns", "schemas", "storage", "store", "table", "where", "primary", "data", "type", "dbindex", "method", "using", "restful", "route") VALUES($1, $2, $3, $4, $5, $6, $7, $8, $9, $10, $11, $12, $13, $14, $15) ON CONFLICT ("key") DO UPDATE SET "datetime"=now(), "foreigns"=$2, "schemas"=$3, "storage"=$4, "store"=$5, "table"=$6, "where"=$7, "primary"=$8, "data"=$9, "type"=$10, "dbindex"=$11, "method"=$12, "using"=$13, "restful"=$14, "route"=$15';
+		executesql := 'INSERT INTO base_redis_foreigns ("key", "foreigns", "schemas", "storage", "store", "table", "where", "primary", "data", "type", "dbindex", "method", "using", "restful", "route") VALUES($1, $2, $3, $4, $5, $6, $7, $8, $9, $10, $11, $12, $13, $14, $15) ON CONFLICT ("key") DO UPDATE SET "datetime"=now(), "foreigns"=$2, "schemas"=$3, "storage"=$4, "store"=$5, "table"=$6, "where"=$7, "primary"=$8, "data"=$9, "type"=$10, "dbindex"=$11, "method"=$12, "using"=$13, "restful"=$14, "route"=$15';
 		EXECUTE executesql USING key_val, foreign_table, schemas_val, storage_val, store_val, table_val, where_val, primary_val, custom_val, type_val, dbindex, method_val, using_val, restful_val, route_val;
 	END CASE;
 	IF(key_val != schemas_val || '_indexedDB_foreigns:list') THEN
